export_zenodo

export_zenodo allows to export a Zenodo record, identified by its DOI or concept DOI, using one of the export formats supported by Zenodo.

export_zenodo(doi, filename, format, append_format = TRUE, logger = NULL)

Arguments

doi

a Zenodo DOI or concept DOI

filename

a base file name (without file extension) to export to.

format

a valid Zenodo export format among the following: BibTeX, CSL, DataCite, DublinCore, DCAT, JSON, JSON-LD, GeoJSON, MARCXML.

append_format

wether format name has to be appended to the filename. Default is TRUE (for backward compatibility reasons). Set it to FALSE if you want to use only the filename.

logger

a logger to print Zenodo API-related messages. The logger can be either NULL, "INFO" (with minimum logs), or "DEBUG" (for complete curl http calls logs)

Value

the exported file name (with extension)

Examples

# \dontrun{
 export_zenodo("10.5281/zenodo.2547036", filename = "test", format = "BibTeX", append_format = F)
#> ✔ Successfully fetched list of published records!
#> ! No record for DOI '10.5281/zenodo.2547036'!
#> ℹ Try to get deposition by Zenodo specific record id '2547036'
#> ✔ Successfully fetched list of published records!
#> ! No record for id '2547036'!
#> ℹ Successfully fetched list of published records - page 1
#> ✔ Successfully fetched list of published records!
#> Error in if (result$conceptdoi == conceptdoi) {    infoMsg = sprintf("Successfully fetched published record for concept DOI '%s'!",         conceptdoi)    cli::cli_alert_success(infoMsg)    self$INFO(infoMsg)} else {    result <- NULL}: argument is of length zero
# }