get_zenodo

get_zenodo allows to get a Zenodo record, identified by its DOI or concept DOI.

get_zenodo(doi, logger = NULL)

Arguments

doi

a Zenodo DOI or concept DOI

logger

a logger to print messages. The logger can be either NULL, "INFO" (with minimum logs), or "DEBUG" (for complete curl http calls logs)

Value

an object of class data.frame giving the record versions including date, version number and version-specific DOI.

object of class ZenodoRecord

Examples

# \dontrun{
 get_zenodo("10.5281/zenodo.2547036")
#> ✔ Successfully fetched list of published records!
#> ! No record for DOI '10.5281/zenodo.2547036'!
#> ℹ Try to get deposition by Zenodo specific record id '2547036'
#> ✔ Successfully fetched list of published records!
#> ! No record for id '2547036'!
#> ℹ Successfully fetched list of published records - page 1
#> ✔ Successfully fetched list of published records!
#> ✔ Successfully fetched published record for concept DOI '10.5281/zenodo.2547036'!
#> <ZenodoRecord>
#> ....|-- access: <NULL>
#> ....|-- conceptdoi: 10.5281/zenodo.2547036
#> ....|-- conceptrecid: 2547036
#> ....|-- created: 2022-06-17T11:23:00.916079+00:00
#> ....|-- doi: 10.5281/zenodo.6656251
#> ....|-- doi_url: https://doi.org/10.5281/zenodo.6656251
#> ....|-- files: 
#> ........1.
#> ........|-- filename: zen4R-0.6-1.tar.gz
#> ........|-- filesize: 51648
#> ........|-- checksum: a7cf6c5a9bda54867ea3bcab2a5dfdf9
#> ........|-- download: https://zenodo.org/api/records/6656251/files/zen4R-0.6-1.tar.gz/content
#> ........2.
#> ........|-- filename: zen4R-0.6-1.zip
#> ........|-- filesize: 65566
#> ........|-- checksum: 0cd3611a83c51e9c23703bd27e1beb92
#> ........|-- download: https://zenodo.org/api/records/6656251/files/zen4R-0.6-1.zip/content
#> ....|-- id: 6656251
#> ....|-- links: 
#> ........|-- self: https://zenodo.org/api/records/6656251
#> ........|-- self_html: https://zenodo.org/records/6656251
#> ........|-- self_doi: https://zenodo.org/doi/10.5281/zenodo.6656251
#> ........|-- doi: https://doi.org/10.5281/zenodo.6656251
#> ........|-- parent: https://zenodo.org/api/records/2547036
#> ........|-- parent_html: https://zenodo.org/records/2547036
#> ........|-- parent_doi: https://zenodo.org/doi/10.5281/zenodo.2547036
#> ........|-- self_iiif_manifest: https://zenodo.org/api/iiif/record:6656251/manifest
#> ........|-- self_iiif_sequence: https://zenodo.org/api/iiif/record:6656251/sequence/default
#> ........|-- files: https://zenodo.org/api/records/6656251/files
#> ........|-- media_files: https://zenodo.org/api/records/6656251/media-files
#> ........|-- archive: https://zenodo.org/api/records/6656251/files-archive
#> ........|-- archive_media: https://zenodo.org/api/records/6656251/media-files-archive
#> ........|-- latest: https://zenodo.org/api/records/6656251/versions/latest
#> ........|-- latest_html: https://zenodo.org/records/6656251/latest
#> ........|-- draft: https://zenodo.org/api/records/6656251/draft
#> ........|-- versions: https://zenodo.org/api/records/6656251/versions
#> ........|-- access_links: https://zenodo.org/api/records/6656251/access/links
#> ........|-- access_users: https://zenodo.org/api/records/6656251/access/users
#> ........|-- access_request: https://zenodo.org/api/records/6656251/access/request
#> ........|-- access: https://zenodo.org/api/records/6656251/access
#> ........|-- reserve_doi: https://zenodo.org/api/records/6656251/draft/pids/doi
#> ........|-- communities: https://zenodo.org/api/records/6656251/communities
#> ........|-- communities-suggestions: https://zenodo.org/api/records/6656251/communities-suggestions
#> ........|-- requests: https://zenodo.org/api/records/6656251/requests
#> ....|-- metadata: 
#> ........|-- title: zen4R: R Interface to Zenodo REST API
#> ........|-- doi: 10.5281/zenodo.6656251
#> ........|-- publication_date: 2022-06-17
#> ........|-- description: <p>zen4R: R Interface to Zenodo REST API</p>
#> 
#> <p><a href="https://github.com/eblondel/zen4R/releases/tag/v0.6-1">https://github.com/eblondel/zen4R/releases/tag/v0.6-1</a></p>
#> ........|-- access_right: open
#> ........|-- creators: 
#> ........|-- contributors: 
#> ........|-- keywords: 
#> ........|-- version: 0.6-1
#> ........|-- language: eng
#> ........|-- resource_type: 
#> ............|-- id: software
#> ........|-- license: 
#> ............|-- id: mit-license
#> ........|-- grants: 
#> ........|-- communities: 
#> ........|-- relations: 
#> ............|-- version: list(index = 12, is_last = FALSE, parent = list(pid_type = "recid", pid_value = "2547036"))
#> ........|-- notes: Sponsors: IRD
#> ....|-- modified: 2023-09-21T06:53:03.924722+00:00
#> ....|-- owners: 
#> ........1.
#> ........|-- id: 5183
#> ....|-- recid: 6656251
#> ....|-- status: published
#> ....|-- state: done
#> ....|-- submitted: TRUE
#> ....|-- revision: 7
#> ....|-- stats:
#>     ↓ downloads = 1195
#>     ↓ unique_downloads = 453
#>     ↓ version_downloads = 41
#>     ↓ version_unique_downloads = 12
#>     👁 views = 1461
#>     👁 unique_views = 1299
#>     👁 version_unique_views = 96
#>     👁 version_views = 98
# }