get_zenodo

get_zenodo allows to get a Zenodo record, identified by its DOI or concept DOI.

get_zenodo(doi, logger = NULL)

Arguments

doi

a Zenodo DOI or concept DOI

logger

a logger to print messages. The logger can be either NULL, "INFO" (with minimum logs), or "DEBUG" (for complete curl http calls logs)

Value

an object of class data.frame giving the record versions including date, version number and version-specific DOI.

object of class ZenodoRecord

Examples

# \dontrun{
 get_zenodo("10.5281/zenodo.2547036")
#> <ZenodoRecord>
#> ....|-- conceptdoi: 10.5281/zenodo.2547036
#> ....|-- conceptrecid: 2547036
#> ....|-- created: 2020-09-02T08:48:36.958406+00:00
#> ....|-- doi: 10.5281/zenodo.4011667
#> ....|-- doi_url: <NULL>
#> ....|-- files: 
#> ........1.
#> ........|-- filename: zen4R-0.4.tar.gz
#> ........|-- filesize: 31849
#> ........|-- checksum: e36420deeaf30ebd25fd74dfdcf6d7cd
#> ........|-- download: https://zenodo.org/api/files/95c3b301-e273-4942-a977-447958e13308/zen4R-0.4.tar.gz
#> ........2.
#> ........|-- filename: zen4R-0.4.zip
#> ........|-- filesize: 41916
#> ........|-- checksum: 5fa140b3837c315bfc9733813b152fd3
#> ........|-- download: https://zenodo.org/api/files/95c3b301-e273-4942-a977-447958e13308/zen4R-0.4.zip
#> ....|-- id: 4011667
#> ....|-- links: 
#> ........|-- badge: https://zenodo.org/badge/doi/10.5281/zenodo.4011667.svg
#> ........|-- bucket: https://zenodo.org/api/files/95c3b301-e273-4942-a977-447958e13308
#> ........|-- conceptbadge: https://zenodo.org/badge/doi/10.5281/zenodo.2547036.svg
#> ........|-- conceptdoi: https://doi.org/10.5281/zenodo.2547036
#> ........|-- doi: https://doi.org/10.5281/zenodo.4011667
#> ........|-- html: https://zenodo.org/record/4011667
#> ........|-- latest: https://zenodo.org/api/records/7552287
#> ........|-- latest_html: https://zenodo.org/record/7552287
#> ........|-- self: https://zenodo.org/api/records/4011667
#> ....|-- metadata: 
#> ........|-- access_right: open
#> ........|-- access_right_category: success
#> ........|-- communities: 
#> ........|-- contributors: 
#> ........|-- creators: 
#> ........|-- description: <p>zen4R: R Interface to Zenodo REST API</p>
#> 
#> <p><a href="https://github.com/eblondel/zen4R/releases/tag/v0.4">https://github.com/eblondel/zen4R/releases/tag/v0.4</a></p>
#> ........|-- doi: 10.5281/zenodo.4011667
#> ........|-- keywords: 
#> ........|-- language: eng
#> ........|-- license: 
#> ............|-- id: MIT
#> ........|-- notes: Sponsors: IRD
#> ........|-- publication_date: 2020-09-02
#> ........|-- related_identifiers: 
#> ........|-- relations: 
#> ............|-- version: list(count = 16, index = 4, is_last = FALSE, last_child = list(pid_type = "recid", pid_value = "7552287"), parent = list(pid_type = "recid", pid_value = "2547036"))
#> ........|-- resource_type: 
#> ............|-- title: Software
#> ............|-- type: software
#> ........|-- title: zen4R: R Interface to Zenodo REST API
#> ........|-- version: 0.4
#> ....|-- modified: <NULL>
#> ....|-- owner: 
#> ........1.
#> ....|-- record_id: <NULL>
#> ....|-- state: <NULL>
#> ....|-- submitted: <NULL>
#> ....|-- title: <NULL>
#> ....|-- version: <NULL>
#> ....|-- stats:
#>     ↓ downloads = 32
#>     ↓ unique_downloads = 6
#>     ↓ version_downloads = 1036
#>     ↓ version_unique_downloads = 359
#>     👁 unique_views = 59
#>     👁 version_unique_views = 851
#>     👁 version_views = 1061
#>     👁 views = 81
#>     ■ version_volume = 36209027
#>     ■ volume = 1180240
# }